+++ react/tests-setup.js
@@ -1,0 +1,9 @@
+
+/* eslint-disable */
+window.matchMedia = window.matchMedia || function () {
+  return {
+    matches: false,
+    addListener: function () { },
+    removeListener: function () { },
+  }
+}