+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.5.0] - 2019-01-16
+### Added
+- Quantity stepper for inline mode
+- Support for products with attachments (for delivery)
+
+### Changed
+- Break index file into smaller components
+
 ## [2.4.1] - 2019-01-14
 ### Fixed
 - Remove `undefined` css classes.